+++ src/store/index.ts
@@ -18,4 +18,4 @@ export {
   resolvePerson
 } from './people'
 export type { PersonCandidate, PersonLinkState, PersonMatch, PersonRef } from './people'
-export { folderOf, projectFilePath, projectFolderOf, projectTaskFolder, TASK_FOLDER_NAME } from './vaultFs'
+export { folderOf, projectFilePath, projectTaskFolder, TASK_FOLDER_NAME } from './vaultFs'